refactor(rwa): move module base surface into compliance modules
Rename the shared module helper file to `storage.rs`, move the shared module
tests to the modules root, and relocate the module-specific trait, errors, and
TTL constants under `rwa::compliance::modules` so the layout matches the
library convention for module namespaces.
