+++ packages/tokens/src/rwa/compliance/modules/mod.rs
@@ -1,6 +1,9 @@
-use soroban_sdk::{contracttrait, Address, Env, String};
+use soroban_sdk::{contracterror, contracttrait, Address, Env, String};
 
-pub mod common;
+pub mod storage;
+
+#[cfg(test)]
+mod test;
 
 /// Trait for compliance modules that can be registered with the modular
 /// compliance system.
@@ -203,3 +206,46 @@ pub trait ComplianceModule {
     /// * `compliance` - The address of the compliance contract.
     fn set_compliance_address(e: &Env, compliance: Address);
 }
+
+// ################## ERRORS ##################
+
+/// Error codes shared by all compliance modules.
+///
+/// Compliance module errors occupy the 390–400 range, following the RWA
+/// error numbering convention.
+#[contracterror]
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[repr(u32)]
+pub enum ComplianceModuleError {
+    /// The compliance contract address has not been set.
+    ComplianceNotSet = 390,
+    /// An amount argument is negative when it must be non-negative.
+    InvalidAmount = 391,
+    /// Arithmetic overflow in a checked addition.
+    MathOverflow = 392,
+    /// Arithmetic underflow in a checked subtraction.
+    MathUnderflow = 393,
+    /// A required limit entry is missing for the given token.
+    MissingLimit = 394,
+    /// A required transfer counter entry is missing.
+    MissingCounter = 395,
+    /// A required country data entry is missing.
+    MissingCountry = 396,
+    /// The identity registry storage address has not been configured.
+    IdentityRegistryNotSet = 397,
+    /// A module is not registered on a required compliance hook.
+    MissingRequiredHook = 398,
+    /// The compliance contract address has already been set.
+    ComplianceAlreadySet = 399,
+    /// A token has reached the maximum number of configured limit entries.
+    TooManyLimits = 400,
+}
+
+// ################## CONSTANTS ##################
+
+const DAY_IN_LEDGERS: u32 = 17280;
+
+/// TTL extension amount for compliance module storage entries (30 days).
+pub const MODULE_EXTEND_AMOUNT: u32 = 30 * DAY_IN_LEDGERS;
+/// TTL threshold below which compliance module entries are extended (29 days).
+pub const MODULE_TTL_THRESHOLD: u32 = MODULE_EXTEND_AMOUNT - DAY_IN_LEDGERS;
